Create bstDriver function which calls tree methods
The bstDriver function will create a random array of numbers to create
a tree. It will then show that the tree is balanced and list the values
in level, pre, post, and in order. Next a few values are added that
unbalance the tree, we show the unbalanced tree, rebalance it, then list
all the values in level, pre, post, and in order again.